+++ rocketpy/sensors/barometer.py
@@ -169,7 +169,6 @@ class Barometer(ScalarSensor):
         pressure = kwargs["environment"].pressure
 
         # Calculate the altitude of the sensor
-
         relative_altitude = (Matrix.transformation(u[6:10]) @ relative_position).z
 
         # Calculate the pressure at the sensor location and add noise