+++ rocketpy/simulation/flight.py
@@ -884,7 +884,7 @@ class Flight:
                 pressure=self.env.pressure,
                 earth_radius=self.env.earth_radius,
                 initial_coordinates=(self.env.latitude, self.env.longitude),
-                rocket=self,
+                rocket=self.rocket,
                 parachute_events=self.parachute_events,
             )